[CSI] Stop endpoint on unpublish request only in vm mode (#3384)

+++ cloud/blockstore/tools/csi_driver/internal/driver/node.go
@@ -1041,38 +1041,40 @@ func (s *nodeService) nodeUnpublishVolume(
 	nbsId := req.VolumeId
 	endpointDir := s.getEndpointDir(podId, nbsId)
 
-	// Trying to stop both NBS and NFS endpoints,
-	// because the endpoint's backend service is unknown here.
-	// When we miss we get S_FALSE/S_ALREADY code (err == nil).
-
-	// Fallback to previous implementation for already mounted volumes to
-	// stop endpoint in nodeUnpublishVolume.
+	// Fallback to previous implementation for already mounted volumes
+	// in VM mode to stop endpoint in nodeUnpublishVolume.
 	// Must be removed after migration of all endpoints to the new format
-	if s.nbsClient != nil {
-		_, err := s.nbsClient.StopEndpoint(ctx, &nbsapi.TStopEndpointRequest{
-			UnixSocketPath: filepath.Join(endpointDir, nbsSocketName),
-		})
-		if err != nil {
-			return fmt.Errorf("failed to stop nbs endpoint: %w", err)
+	if s.vmMode {
+		// Trying to stop both NBS and NFS endpoints,
+		// because the endpoint's backend service is unknown here.
+		// When we miss we get S_FALSE/S_ALREADY code (err == nil).
+		if s.nbsClient != nil {
+			_, err := s.nbsClient.StopEndpoint(ctx, &nbsapi.TStopEndpointRequest{
+				UnixSocketPath: filepath.Join(endpointDir, nbsSocketName),
+			})
+			if err != nil {
+				return fmt.Errorf("failed to stop nbs endpoint: %w", err)
+			}
 		}
-	}
 
-	nfsClient := s.getNfsClient(nbsId)
-	if nfsClient != nil {
-		_, err := nfsClient.StopEndpoint(ctx, &nfsapi.TStopEndpointRequest{
-			SocketPath: filepath.Join(endpointDir, nfsSocketName),
-		})
-		if err != nil {
-			return fmt.Errorf("failed to stop nfs endpoint (%T): %w", nfsClient, err)
+		nfsClient := s.getNfsClient(nbsId)
+		if nfsClient != nil {
+			_, err := nfsClient.StopEndpoint(ctx, &nfsapi.TStopEndpointRequest{
+				SocketPath: filepath.Join(endpointDir, nfsSocketName),
+			})
+			if err != nil {
+				return fmt.Errorf("failed to stop nfs endpoint (%T): %w", nfsClient, err)
+			}
 		}
+
+		if err := os.RemoveAll(endpointDir); err != nil {
+			return err
+		}
+
+		// remove pod's folder if it's empty
+		ignoreError(os.Remove(s.getEndpointDir(podId, "")))
 	}
 
-	if err := os.RemoveAll(endpointDir); err != nil {
-		return err
-	}
-
-	// remove pod's folder if it's empty
-	ignoreError(os.Remove(s.getEndpointDir(podId, "")))
 	return nil
 }
